feat(5.15): add video hook
